Restructure resolver-map module
